test(server): update test
